requirements.txt
# Celery
celery==4.4.5
kombu==4.6.11
redis==3.5.3
billiard==3.6.3.0
flower==0.9.5

# Dev
ipython==7.16.1
pytest==6.0.1
python-dotenv==0.14.0
